Move test data providers
